+++ HarmonyDB.Index/HarmonyDB.Index.Analysis/Models/Index/StructureLoop.cs
@@ -1,3 +1,3 @@
 ﻿namespace HarmonyDB.Index.Analysis.Models.Index;
 
-public record StructureLoop(string Normalized, int Length, int TotalOccurrences, int TotalSuccessions, int TotalSongs);
+public record StructureLoop(string Normalized, int Length, float TotalOccurrences, float TotalSuccessions, int TotalSongs);
